feat(sdk)!: split plugin UI (Content) into a separate interface
Make headless plugins possible by removing the UI from the base plugin types and
moving it to opt-in UI types:

- JetWhaleRawHostPlugin (base) no longer declares ContentRaw; it is now headless.
- New JetWhaleRawUiHostPlugin interface carries the @Composable ContentRaw; the
  host renders a scene only for plugins that implement it.
- JetWhaleHostPlugin<E,M,R> (typed base) is now headless (event handling only).
- New JetWhaleUiHostPlugin<E,M,R> adds the typed @Composable Content and the
  raw->typed context adapter; extend it for plugins with a UI.

Host: DefaultPluginComposeSceneService renders ContentRaw only when the instance
is a JetWhaleRawUiHostPlugin. In-repo example/network plugins extend
JetWhaleUiHostPlugin.

BREAKING CHANGE: plugins with a UI must extend JetWhaleUiHostPlugin instead of
JetWhaleHostPlugin (or implement JetWhaleRawUiHostPlugin alongside
JetWhaleRawHostPlugin).
