+++ jetwhale-host-sdk/api/jetwhale-host-sdk.api
@@ -9,8 +9,6 @@ public abstract interface class com/kitakkun/jetwhale/host/sdk/JetWhaleDebugOper
 public abstract class com/kitakkun/jetwhale/host/sdk/JetWhaleHostPlugin : com/kitakkun/jetwhale/host/sdk/JetWhaleRawHostPlugin {
 	public static final field $stable I
 	public fun <init> ()V
-	public abstract fun Content (Lcom/kitakkun/jetwhale/host/sdk/JetWhaleDebugOperationContext;Landroidx/compose/runtime/Composer;I)V
-	public final fun ContentRaw (Lcom/kitakkun/jetwhale/host/sdk/JetWhaleRawDebugOperationContext;Landroidx/compose/runtime/Composer;I)V
 	protected abstract fun getProtocol ()Lcom/kitakkun/jetwhale/protocol/host/JetWhaleHostPluginProtocol;
 	public abstract fun onEvent (Ljava/lang/Object;)V
 	public final fun onRawEvent (Ljava/lang/String;)V
@@ -202,8 +200,18 @@ public abstract interface class com/kitakkun/jetwhale/host/sdk/JetWhaleRawDebugO
 public abstract class com/kitakkun/jetwhale/host/sdk/JetWhaleRawHostPlugin {
 	public static final field $stable I
 	public fun <init> ()V
-	public abstract fun ContentRaw (Lcom/kitakkun/jetwhale/host/sdk/JetWhaleRawDebugOperationContext;Landroidx/compose/runtime/Composer;I)V
 	public fun onDispose ()V
 	public abstract fun onRawEvent (Ljava/lang/String;)V
+}
+
+public abstract interface class com/kitakkun/jetwhale/host/sdk/JetWhaleRawUiHostPlugin {
+	public abstract fun ContentRaw (Lcom/kitakkun/jetwhale/host/sdk/JetWhaleRawDebugOperationContext;Landroidx/compose/runtime/Composer;I)V
+}
+
+public abstract class com/kitakkun/jetwhale/host/sdk/JetWhaleUiHostPlugin : com/kitakkun/jetwhale/host/sdk/JetWhaleHostPlugin, com/kitakkun/jetwhale/host/sdk/JetWhaleRawUiHostPlugin {
+	public static final field $stable I
+	public fun <init> ()V
+	public abstract fun Content (Lcom/kitakkun/jetwhale/host/sdk/JetWhaleDebugOperationContext;Landroidx/compose/runtime/Composer;I)V
+	public final fun ContentRaw (Lcom/kitakkun/jetwhale/host/sdk/JetWhaleRawDebugOperationContext;Landroidx/compose/runtime/Composer;I)V
 }
